requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.1.2
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.2.0
pillow==11.3.0
prettytable==3.16.0
python-decouple==3.8
python-http-client==3.3.7
pytz==2023.3.post1
reportlab==4.4.3
requests==2.31.0
sendgrid==6.10.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
typing_extensions==4.7.1
urllib3==2.0.7
wcwidth==0.2.13
